// src/app/models/person.interface.ts
export interface PersonMinimal {
  uid: string;
  name: string;
  url: string;
}

export interface PersonDetailed {
  height: string;
  mass: string;
  hair_color: string;
  skin_color: string;
  eye_color: string;
  birth_year: string;
  gender: string;
  created: string;
  edited: string;
  name: string;
  homeworld: string;
  url: string;
}